feat: refactor charts, add insights, configure linting
- Move renderHeatmap/renderPieChart/renderTrendChart to charts.js
- Add insight cards (racha, días sin cardio/fuerza, tendencia de pasos)
- Configure ESLint 9 + Prettier with lint/format npm scripts
- Bump SW cache to v3 to invalidate previous cache
