+++ docs/css/style.css
@@ -132,6 +132,79 @@ body {
   text-transform: uppercase;
 }
 
+/* ── INSIGHTS ── */
+.insights-section {
+  background: linear-gradient(135deg, rgba(30, 41, 59, 0.85), rgba(17, 24, 39, 0.8));
+  border: 1px solid rgba(255, 255, 255, 0.06);
+  backdrop-filter: blur(8px);
+  border-radius: 14px;
+  padding: 28px;
+  margin-bottom: 24px;
+  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.3);
+}
+
+.insights-section h2 {
+  font-size: 1.05em;
+  font-weight: 700;
+  color: #e2e8f0;
+  margin-bottom: 20px;
+  text-transform: uppercase;
+  letter-spacing: 0.06em;
+  border-left: 3px solid #a78bfa;
+  padding-left: 12px;
+}
+
+.insights-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(240px, 1fr));
+  gap: 12px;
+}
+
+.insight-card {
+  display: flex;
+  align-items: flex-start;
+  gap: 14px;
+  background: rgba(15, 23, 42, 0.55);
+  border: 1px solid rgba(255, 255, 255, 0.06);
+  border-radius: 10px;
+  padding: 16px;
+  transition: transform 0.2s, border-color 0.2s;
+}
+
+.insight-card:hover {
+  transform: translateY(-2px);
+}
+
+.insight-card.insight-positive { border-left: 3px solid #34d399; }
+.insight-card.insight-warning  { border-left: 3px solid #fb923c; }
+.insight-card.insight-neutral  { border-left: 3px solid #64748b; }
+
+.insight-icon {
+  font-size: 1.6em;
+  line-height: 1;
+  flex-shrink: 0;
+  margin-top: 2px;
+}
+
+.insight-body {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.insight-title {
+  font-family: 'Roboto Mono', monospace;
+  font-size: 0.88em;
+  font-weight: 700;
+  color: #e2e8f0;
+}
+
+.insight-desc {
+  font-size: 0.78em;
+  color: #6b7280;
+  line-height: 1.4;
+}
+
 /* ── RACHA RECIENTE ── */
 .racha-section {
   background: linear-gradient(135deg, rgba(30, 41, 59, 0.85), rgba(17, 24, 39, 0.8));
