+++ docs/index.html
@@ -42,6 +42,11 @@
       </div>
     </section>
 
+    <section class="insights-section">
+      <h2>Insights</h2>
+      <div class="insights-grid" id="insightsGrid"></div>
+    </section>
+
     <section class="racha-section">
       <h2>Racha reciente</h2>
       <div class="racha-cards" id="rachaCards"></div>
