+++ docs/js/main.js
@@ -1,3 +1,5 @@
+/* global renderCharts */
+
 let dashboardData = null;
 
 async function loadData() {
@@ -45,6 +47,9 @@ function renderDashboard() {
   document.getElementById('diasEjercicio').textContent = stats.diasEjercicio;
   document.getElementById('diasPasos').textContent = stats.diasPasos;
   document.getElementById('promedioPasos').textContent = stats.promedioPasos.toLocaleString('es-MX');
+
+  // Insights automáticos
+  renderInsights();
 
   // Racha reciente
   renderRachas();
@@ -108,7 +113,7 @@ function renderWeekTable() {
   const diasNombre = ['Domingo', 'Lunes', 'Martes', 'Miércoles', 'Jueves', 'Viernes', 'Sábado'];
   const porDiaSemana = dashboardData.porDiaSemana;
 
-  diasNombre.forEach((dia, idx) => {
+  diasNombre.forEach((dia, _idx) => {
     const stats = porDiaSemana[dia.substring(0, 3)] || { total: 0, ejercicio: 0, pasos: 0, verdes: 0 };
     
     const pctEjercicio = stats.total > 0 ? Math.round((stats.ejercicio / stats.total) * 100) : 0;
@@ -222,300 +227,94 @@ function renderWeeklySummary() {
   }).join('');
 }
 
-function renderCharts() {
-  renderHeatmap();
-  renderPieChart();
-  renderTrendChart();
+function computeInsights() {
+  const datos = dashboardData.todosDatos;
+  const hoy = new Date().toISOString().split('T')[0];
+  const sorted = [...datos].sort((a, b) => b.fecha.localeCompare(a.fecha));
+
+  let racha = 0;
+  for (const d of sorted) {
+    if (d.fecha > hoy) continue;
+    if (d.estado === 'verde' || d.estado === 'amarillo') racha++;
+    else break;
+  }
+
+  const CARDIO = ['Natación', 'Carrera', 'Bici', 'Mixto'];
+  const ultimoCardio = sorted.find(d => d.fecha <= hoy && d.disciplinas.some(disc => CARDIO.includes(disc)));
+  const diasSinCardio = ultimoCardio
+    ? Math.round((new Date(hoy) - new Date(ultimoCardio.fecha)) / 86400000)
+    : null;
+
+  const ultimoFuerza = sorted.find(d => d.fecha <= hoy && (d.disciplinas.includes('Fuerza') || d.disciplinas.includes('Mixto')));
+  const diasSinFuerza = ultimoFuerza
+    ? Math.round((new Date(hoy) - new Date(ultimoFuerza.fecha)) / 86400000)
+    : null;
+
+  const hace7 = new Date();
+  hace7.setDate(hace7.getDate() - 6);
+  const hace7Str = hace7.toISOString().split('T')[0];
+  const ultimos7 = datos.filter(d => d.fecha >= hace7Str && d.fecha <= hoy && d.pasos > 0);
+  const promedioUltimos7 = ultimos7.length
+    ? Math.round(ultimos7.reduce((s, d) => s + d.pasos, 0) / ultimos7.length)
+    : 0;
+  const diffPasos = dashboardData.stats.promedioPasos > 0
+    ? Math.round(((promedioUltimos7 - dashboardData.stats.promedioPasos) / dashboardData.stats.promedioPasos) * 100)
+    : 0;
+
+  return { racha, diasSinCardio, diasSinFuerza, promedioUltimos7, diffPasos };
 }
 
-function renderTrendChart() {
-  const ctx = document.getElementById('trendChart');
-  if (!ctx || !dashboardData.todosDatos) return;
-  
-  const todosLosDatos = dashboardData.todosDatos;
-  const labels = todosLosDatos.map(d => {
-    const date = new Date(d.fecha + 'T00:00:00');
-    return date.toLocaleDateString('es-MX', { day: 'numeric', month: 'short' });
-  });
-  const data = todosLosDatos.map(d => d.pasos);
+function renderInsights() {
+  const container = document.getElementById('insightsGrid');
+  if (!container) return;
 
-  new Chart(ctx.getContext('2d'), {
-    type: 'line',
-    data: {
-      labels,
-      datasets: [
-        {
-          label: 'Pasos diarios',
-          data,
-          borderColor: '#00d4ff',
-          backgroundColor: 'rgba(0, 212, 255, 0.05)',
-          fill: true,
-          tension: 0.2,
-          pointRadius: todosLosDatos.length > 60 ? 1 : 3,
-          pointHoverRadius: 6,
-          borderWidth: 2,
-          pointBackgroundColor: '#00d4ff',
-        },
-        {
-          label: 'Meta (10k)',
-          data: Array(labels.length).fill(10000),
-          borderColor: 'rgba(52, 211, 153, 0.4)',
-          borderDash: [5, 5],
-          fill: false,
-          pointRadius: 0,
-          borderWidth: 1.5,
-        }
-      ]
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      interaction: {
-        mode: 'index',
-        intersect: false,
-      },
-      plugins: {
-        legend: {
-          display: true,
-          position: 'top',
-          labels: { 
-            color: '#9ca3af',
-            boxWidth: 20,
-            font: { size: 11 }
-          }
-        },
-        tooltip: {
-          backgroundColor: 'rgba(15, 23, 42, 0.9)',
-          padding: 12,
-          callbacks: {
-            label: function(context) {
-              const val = context.parsed.y;
-              return `${context.dataset.label}: ${val ? val.toLocaleString('es-MX') : 0}`;
-            }
-          }
-        }
-      },
-      scales: {
-        y: {
-          beginAtZero: true,
-          grid: { color: 'rgba(255, 255, 255, 0.05)' },
-          ticks: { 
-            color: '#64748b',
-            callback: function(value) {
-              return value >= 1000 ? (value / 1000) + 'k' : value;
-            }
-          }
-        },
-        x: {
-          grid: { display: false },
-          ticks: { 
-            color: '#64748b',
-            maxRotation: 45,
-            minRotation: 45,
-            autoSkip: true,
-            maxTicksLimit: 15
-          }
-        }
-      }
-    }
-  });
-}
+  const { racha, diasSinCardio, diasSinFuerza, promedioUltimos7, diffPasos } = computeInsights();
+  const insights = [];
 
-function renderHeatmap() {
-  const container = document.getElementById('heatmap');
-  container.innerHTML = '';
+  if (racha >= 2) {
+    insights.push({ icon: '🔥', title: `${racha} días seguidos`, desc: `Llevas ${racha} días consecutivos cumpliendo el objetivo.`, type: 'positive' });
+  } else {
+    insights.push({ icon: '⚠️', title: 'Sin racha activa', desc: 'No hay cumplimiento consecutivo registrado. ¡Momento de retomar!', type: 'warning' });
+  }
 
-  const colorMap = {
-    verde:   '#34d399',
-    amarillo: '#fb923c',
-    rojo:    '#f87171',
-    gris:    '#1e293b',
-    futuro:  'transparent'
-  };
-  const estadoLabel = { verde: 'Verde', amarillo: 'Naranja', rojo: 'Rojo', gris: 'Sin datos', futuro: 'Próximamente' };
-  const MONTHS = ['Ene','Feb','Mar','Abr','May','Jun','Jul','Ago','Sep','Oct','Nov','Dic'];
-
-  const hoy = new Date();
-  
-  // Iniciar exactamente el 29 de diciembre de 2025
-  const startYear = new Date('2025-12-29T00:00:00');
-  // Terminar al final del 2026
-  const endYear = new Date(2026, 11, 31, 23, 59, 59);
-  
-  const diasMap = Object.fromEntries(dashboardData.todosLosDias.map(d => [d.fecha, d]));
-  
-  const diasCompletos = [];
-  for (let d = new Date(startYear); d <= endYear; d.setDate(d.getDate() + 1)) {
-    const fStr = d.toISOString().split('T')[0];
-    if (diasMap[fStr]) {
-      // Si el día está en el mapa pero es después de hoy, forzamos estado futuro
-      const dia = {...diasMap[fStr]};
-      if (d > hoy) dia.estado = 'futuro';
-      diasCompletos.push(dia);
+  if (diasSinCardio !== null) {
+    if (diasSinCardio === 0) {
+      insights.push({ icon: '🏊', title: 'Cardio hoy', desc: 'Hiciste cardio hoy.', type: 'positive' });
+    } else if (diasSinCardio > 4) {
+      insights.push({ icon: '🏊', title: `${diasSinCardio} días sin cardio`, desc: `La última sesión de cardio fue hace ${diasSinCardio} días.`, type: 'warning' });
     } else {
-      diasCompletos.push({
-        fecha: fStr,
-        estado: d > hoy ? 'futuro' : 'gris',
-        pasos: 0, disciplinas: [], duracion: 0
-      });
+      insights.push({ icon: '🏊', title: `Cardio hace ${diasSinCardio}d`, desc: `Última sesión de cardio hace ${diasSinCardio} día${diasSinCardio > 1 ? 's' : ''}.`, type: 'neutral' });
     }
   }
 
-  // Pad start so first day aligns to its day-of-week (Mon = row 0)
-  const startPad  = (startYear.getDay() + 6) % 7;
-  const cells     = [...Array(startPad).fill(null), ...diasCompletos];
-  const totalWeeks = Math.ceil(cells.length / 7);
-
-  // ── Month labels ──────────────────────────────────────────
-  const monthRow = document.createElement('div');
-  monthRow.className = 'heatmap-months';
-
-  let lastMonth = -1;
-  for (let w = 0; w < totalWeeks; w++) {
-    const span = document.createElement('span');
-    for (let r = 0; r < 7; r++) {
-      const c = cells[w * 7 + r];
-      if (c) {
-        const m = new Date(c.fecha + 'T00:00:00').getMonth();
-        if (m !== lastMonth) { 
-          span.textContent = MONTHS[m]; 
-          lastMonth = m; 
-        }
-        break;
-      }
-    }
-    monthRow.appendChild(span);
-  }
-
-  // ── Grid row (day labels + cells) ────────────────────────
-  const gridRow = document.createElement('div');
-  gridRow.className = 'heatmap-grid-row';
-
-  const dayLabels = document.createElement('div');
-  dayLabels.className = 'heatmap-day-labels';
-  ['L','M','X','J','V','S','D'].forEach(lbl => {
-    const el = document.createElement('span');
-    el.textContent = lbl;
-    dayLabels.appendChild(el);
-  });
-
-  const grid = document.createElement('div');
-  grid.className = 'heatmap-grid';
-
-  // Shared tooltip
-  let tip = document.getElementById('heatmap-tip');
-  if (!tip) {
-    tip = document.createElement('div');
-    tip.id = 'heatmap-tip';
-    tip.className = 'heatmap-tooltip';
-    document.body.appendChild(tip);
-  }
-
-  cells.forEach(dia => {
-    const cell = document.createElement('div');
-    cell.className = 'heatmap-cell';
-
-    if (dia) {
-      if (dia.estado === 'futuro') {
-        cell.classList.add('futuro');
-      }
-      
-      const color = colorMap[dia.estado];
-      cell.style.background = color;
-      
-      if (dia.estado !== 'gris' && dia.estado !== 'futuro') {
-        cell.style.boxShadow = `0 0 5px ${color}55`;
-      }
-
-      cell.addEventListener('mouseenter', e => {
-        tip.innerHTML =
-          `<div class="tt-date">${dia.fecha}</div>` +
-          `<div class="tt-row">Estado: <strong>${estadoLabel[dia.estado]}</strong></div>` +
-          (dia.pasos > 0 ? `<div class="tt-row">Pasos: <strong>${dia.pasos.toLocaleString()}</strong></div>` : '') +
-          (dia.disciplinaPrincipal ? `<div class="tt-row">Disciplina: <strong>${dia.disciplinaPrincipal}</strong></div>` : '') +
-          (dia.duracion ? `<div class="tt-row">Duración: <strong>${dia.duracion} min</strong></div>` : '');
-        tip.classList.add('visible');
-        moveTip(e);
-      });
-      cell.addEventListener('mousemove', moveTip);
-      cell.addEventListener('mouseleave', () => tip.classList.remove('visible'));
+  if (diasSinFuerza !== null) {
+    if (diasSinFuerza === 0) {
+      insights.push({ icon: '💪', title: 'Fuerza hoy', desc: 'Hiciste una sesión de fuerza hoy.', type: 'positive' });
+    } else if (diasSinFuerza > 3) {
+      insights.push({ icon: '💪', title: `${diasSinFuerza} días sin fuerza`, desc: `La última sesión de fuerza fue hace ${diasSinFuerza} días.`, type: 'warning' });
     } else {
-      cell.style.background = 'transparent';
-      cell.style.pointerEvents = 'none';
+      insights.push({ icon: '💪', title: `Fuerza hace ${diasSinFuerza}d`, desc: `Última sesión de fuerza hace ${diasSinFuerza} día${diasSinFuerza > 1 ? 's' : ''}.`, type: 'neutral' });
     }
-    grid.appendChild(cell);
-  });
-
-  function moveTip(e) {
-    tip.style.left = (e.clientX + 14) + 'px';
-    tip.style.top  = (e.clientY - 10) + 'px';
   }
 
-  gridRow.appendChild(dayLabels);
-  gridRow.appendChild(grid);
-  container.appendChild(monthRow);
-  container.appendChild(gridRow);
-}
+  if (promedioUltimos7 > 0) {
+    const signo = diffPasos >= 0 ? '+' : '';
+    insights.push({
+      icon: '👟',
+      title: `${promedioUltimos7.toLocaleString('es-MX')} pasos/día`,
+      desc: `Promedio últimos 7 días: ${signo}${diffPasos}% vs. tu media general (${dashboardData.stats.promedioPasos.toLocaleString('es-MX')}).`,
+      type: diffPasos >= 0 ? 'positive' : 'warning',
+    });
+  }
 
-function renderPieChart() {
-  const ctx = document.getElementById('pieChart').getContext('2d');
-
-  const disciplinas = dashboardData.distribucionDisciplinas;
-  const labels = disciplinas.map(d => `${d.nombre}: ${d.count} (${d.porcentaje}%)`);
-  const data = disciplinas.map(d => d.count);
-
-  const colores = [
-    '#00d4ff',
-    '#34d399',
-    '#a78bfa',
-    '#fbbf24',
-    '#f87171',
-    '#fb923c',
-  ];
-
-  new Chart(ctx, {
-    type: 'doughnut',
-    data: {
-      labels,
-      datasets: [
-        {
-          data,
-          backgroundColor: colores.slice(0, data.length),
-          borderColor: 'rgba(15, 23, 42, 0.9)',
-          borderWidth: 2,
-        },
-      ],
-    },
-    options: {
-      responsive: true,
-      maintainAspectRatio: true,
-      plugins: {
-        legend: {
-          position: 'bottom',
-          labels: {
-            padding: 16,
-            color: '#9ca3af',
-            font: { family: 'Roboto', size: 12 },
-            usePointStyle: true,
-            pointStyleWidth: 10,
-          },
-        },
-        tooltip: {
-          backgroundColor: 'rgba(15, 23, 42, 0.95)',
-          titleColor: '#00d4ff',
-          bodyColor: '#cbd5e1',
-          borderColor: 'rgba(0, 212, 255, 0.2)',
-          borderWidth: 1,
-          padding: 12,
-          callbacks: {
-            label: function(context) {
-              return `${context.label}: ${context.parsed} sesiones`;
-            },
-          },
-        },
-      },
-    },
-  });
+  container.innerHTML = insights.map(ins => `
+    <div class="insight-card insight-${ins.type}">
+      <span class="insight-icon">${ins.icon}</span>
+      <div class="insight-body">
+        <div class="insight-title">${ins.title}</div>
+        <div class="insight-desc">${ins.desc}</div>
+      </div>
+    </div>`).join('');
 }
 
 // Cargar datos al iniciar
